scripts/requirements.txt
pyserial>=3.4
esptool>=4.0 